+++ CustomerGreen.Core/Entities/FeedbackCategory.cs
@@ -2,7 +2,7 @@
 {
     public class FeedbackCategory : BaseEntity
     {
-        public int OrgId { get; set; }
+        public long OrgId { get; set; }
         public string CategoryKey { get; set; }
         public string Category { get; set; }
     }
